Minimum version to start with
Add the C++ source and header files for the CLI test application and the dynamic library. This works fine.
Add the SWIG interface file for Python wrapping. Not working yet because of include issues (spdlog at least).
Add the setup.py script to create the Python package. Not tested yet.